requirements.txt
argparse>=1.2.1
beautifulsoup4>=4.0.2
coverage>=3.5.2
Mako>=0.7.0
Markdown>=2.1.1
nose>=1.1.2
pep8>=1.3.3
pylint>=0.25.2
PyYAML>=3.10
Sphinx>=1.1.3
